refactor(codegen): extract compile_block into control_flow.rs (#113 phase 4)
