+++ src/codegen/mod.rs
@@ -1,4 +1,5 @@
 pub mod compiler;
+mod control_flow;
 mod generics;
 mod intrinsics;
 pub mod transpiler;
